Split epochs module from stim features
