+++ tests/test_stim_features.py
@@ -2,33 +2,6 @@ import ipfx.stim_features as st
 import numpy as np
 
 
-def test_find_stim_window():
-    a = [0,1,1,0]
-    stim_start, stim_dur = st.find_stim_window(a)
-    assert stim_start == 1
-    assert stim_dur == 2
-
-    a = [0,1,1,0,1,1,0]
-    stim_start, stim_dur = st.find_stim_window(a, idx0=3)
-    assert stim_start == 4
-    assert stim_dur == 2
-
-    a = [1,1,0,0,0,0,0]
-    stim_start, stim_dur = st.find_stim_window(a)
-    assert stim_start == 2
-    assert stim_dur == 5
-
-    a = [1,1,0,0,0,0,1]
-    stim_start, stim_dur = st.find_stim_window(a)
-    assert stim_start == 2
-    assert stim_dur == 4
-
-    a = [1,1,0,0,1,0,1]
-    stim_start, stim_dur = st.find_stim_window(a)
-    assert stim_start == 2
-    assert stim_dur == 4
-
-
 def test_get_stim_characteristics():
 
     i = [0,1,1,0]
@@ -111,6 +84,7 @@ def test_get_stim_characteristics():
     assert np.isclose(amplitude, 0.7)
     assert start_idx == 6
     assert end_idx == 12
+
 
 def test_find_stim_interval():
     a = [0,1,0,1,0,1,0]
